+++ src/Utils/EndianUtils.h
@@ -23,8 +23,6 @@
 #ifndef _EndianUtils_h_
 #define _EndianUtils_h_
 
-#include <stdint.h>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -74,8 +72,8 @@ PlatformType	GetNativePlatformType(void);
  *
  */
 
-void	EndianFlipShort(int16_t * ioShort);
-void	EndianFlipLong(int32_t * ioLong);
+void	EndianFlipShort(short * ioShort);
+void	EndianFlipLong(long * ioLong);
